+++ src/Security/Security/Models/IotSecuritySolutionAnalytics/PSIoTSecurityDeviceAlert.cs
@@ -20,6 +20,6 @@ namespace Microsoft.Azure.Commands.Security.Models.IotSecuritySolutionAnalytics
 
         public string ReportedSeverity { get; set; }
 
-        public long? AlertsCount { get; set; }
+        public int? AlertsCount { get; set; }
     }
 }
